+++ docs/content/user/changelog.md
@@ -6,6 +6,7 @@ weight: 100
 ## swanshell 1.1.2 (Eventually)
 
 - Added: Long file names are now automatically scrolled in the file selector.
+- Added: New USB shell commands: `cat`, `echo`, `pwd`.
 - Fixed: Regression in swanshell 1.1.0 which led to slower XMODEM transfer start times.
 
 ## swanshell 1.1.1 (4th March 2026)